requirements.txt
python-telegram-bot==21.3
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1